Fix memory leak
The problem was caused by trying to deal with _optionally_ compressed
data without copying it: data segments which are not compressed are
simply a slice of the main buffer, and should not be freed by
themselves. Data segments which must be decompressed need to be
allocated some memory to decompress into, and that must be freed after
use.

Solution was to use a wrapper struct to handle this, which records if
the data should be freed or not, and does the free inside it's deinit()
method.