Return a single error type, providing more detailed reason for failure in an optional Diagnostics structure.
This is an idomatic approach in Zig, which allows API users to decide
the granularity of error information they receive.

This initial conversion simply moves the existing error set to an enum
and sets it as a subtype member on the new Diagnostics structure. The
diagnostics structure could be extended in future with human readable
errors, or further detail on the position of errors in the file etc.